separated controller and view
